finish all homework

+++ DataStructure/Homework/HW3/F74092269_hw3-4.cpp
@@ -5,30 +5,36 @@
 
 using namespace std;
 
+// Edge to record the adjacent vertices and cost between them
 struct Edge{
     int v1, v2;
     int cost;
 };
 
+// Myset is used to record the connected component
 class MySet{
     private:
-        list <int> elements;
+        list <int> elements; // connected component
     public:
         void add(int value){
+            // add to list buttom
             this->elements.push_back(value);
         }
     friend class MySets;
 };
 
+// record how many sets are their and each component
 class MySets{
     private:
         int numOfSets;
-        MySet* sets;
+        MySet* sets; // components
     public:
         MySets(int numOfSets){
+            // first initialize the set's number to numOfVertex
             this->numOfSets = numOfSets;
             this->sets = new MySet[numOfSets];
-
+            // let all vertex construct one set1
+            // therefore totally, there are numOfVertex sets
             for (int i = 0; i < numOfSets; i++){
                 this->sets[i].add(i);
             }
@@ -39,19 +45,23 @@ class MySets{
 
 int main(){
 
-    ifstream ifs;
-    ofstream ofs;
+    ifstream ifs; // for input file
+    ofstream ofs; // for output file
 
+    // input the file path and read it
     string inputpath;
     cout << "FILEPATH: ";
     getline(cin, inputpath);
     ifs.open(inputpath);
 
     if (!ifs.is_open()){
+        // error to open it
         cout << "Failed to open file." << endl;
         return 1;
     }
 
+    // replace 'in' string to 'out' string
+    // and open the output file
     int found = inputpath.find("in");
     string outputpath = inputpath.replace(found, 2, "out");
     ofs.open(outputpath);
@@ -59,16 +69,17 @@ int main(){
     int numOfVertex, numOfEdge;
     ifs >> numOfVertex >> numOfEdge;
 
-    int v1, v2, cost;
-    long long int totalCost = 0;
-    int totalEdges = 0;
-    Edge* edgeSet = new Edge[numOfEdge];
-    MySets vertexSet(numOfVertex);
+    int v1, v2, cost; // two adjacent vertices and the cost between them
+    long long int totalCost = 0; // record the total cost (long long int in case overflow)
+    int totalEdges = 0; // record how many edges is choosen
+    Edge* edgeSet = new Edge[numOfEdge]; // edges' array
+    MySets vertexSet(numOfVertex); // initialize
 
     for (int i = 0; i < numOfEdge; i++){
         ifs >> v1 >> v2 >> cost;
-        edgeSet[i] = {v1, v2, cost};
+        edgeSet[i] = {v1, v2, cost}; // construct the edges' array
     }
+
     // bubble sort from smaller one to bigger one
     for (int i = 0; i < numOfEdge; i++){
         for (int j = 0; j < numOfEdge - 1 - i; j++){
@@ -82,17 +93,22 @@ int main(){
         }
     }
 
+    // now edge array is in non decreasing order
+
     for (int i = 0; i < numOfEdge && totalEdges < numOfVertex - 1; i++){
+        // go through edge array and
+        // finding the set the vertices from
         int setIdx1 = vertexSet.Find(edgeSet[i].v1);
         int setIdx2 = vertexSet.Find(edgeSet[i].v2);
 
         if (setIdx1 == setIdx2){
+            // in the same set
             // form a cycle and discard this edge
             continue;
         }else{
             totalEdges += 1;
-            totalCost += edgeSet[i].cost;
-            vertexSet.Union(setIdx1, setIdx2);
+            totalCost += (long long int)edgeSet[i].cost;
+            vertexSet.Union(setIdx1, setIdx2); // Union the two set for next usage
         }
     }
 
@@ -105,10 +121,14 @@ int main(){
     return 0;
 }
 
+// find the set index for specific vertex
 int MySets::Find(int value){
     for (int i = 0; i < this->numOfSets; i++){
+        // go through the sets
         for (list<int>::iterator it = this->sets[i].elements.begin(); it != this->sets[i].elements.end(); it++){
+            // go through set's elements
             if (value == *it){
+                // find out and return
                 return i;
             }
         }
@@ -116,8 +136,9 @@ int MySets::Find(int value){
     return -1; // discard the warning because finding function must find
 }
 
+// Union two sets
 void MySets::Union(int setIndex1, int setIndex2){
-    bool formerMax = (setIndex1 >= setIndex2);
+    bool formerMax = (setIndex1 >= setIndex2); // check the former is bigger or not
 
     if (formerMax){
         // set2 add to set1
